fix:searchdialog own component
